% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_bindings}
\alias{generate_bindings}
\title{Generate Bindings from GIR files}
\usage{
generate_bindings(gir_paths, out_dir)
}
\arguments{
\item{gir_paths}{Character vector of paths to GIR files}

\item{out_dir}{Directory containing R/ and src/ subdirectories}
}
\description{
Generate Bindings from GIR files
}
